feat: add custom identity template builder
