--- .github/workflows/apply-custom-template-ui.yml
@@ -1,35 +1,0 @@
-name: Apply custom template UI
-
-on:
-  push:
-    branches:
-      - feat/v2.5.7-custom-templates
-    paths:
-      - .github/workflows/apply-custom-template-ui.yml
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-        with:
-          ref: feat/v2.5.7-custom-templates
-          fetch-depth: 0
-
-      - name: Apply and verify
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 .agent/apply_custom_template.py
-          node --check module/template/webroot/policy.js
-          node --check module/template/webroot/ux.js
-          for test in module/webui-tests/*.test.js; do node "$test"; done
-          git config user.name "Yiğit"
-          git config user.email "40565628+tryigit@users.noreply.github.com"
-          git add -A
-          git commit -m "feat: add custom identity template builder"
-          git push origin HEAD:feat/v2.5.7-custom-templates
