--- docs/.trusted-write-check
@@ -1,1 +1,0 @@
-trusted-access write check
